Created Data Folder and Game Repository Interface
Created a Data folder for storing object repository references and functions.

The Game Repository Interface was added to this folder which allows for loose coupling of classes with the interfaces methods and enforces the interface methods used by those who implement them.

A Default Game Repository class was added to the Data folder implementing the Game Repository interface. This is primarily for testing and seeing basic implementation of the base functions of the interface.