Endpoint 1 Request Functionality Completed
Endpoint 1, which provides functionality to post a newly created game and its players has been developed at the request level. This functionality has been tested using Postman and the DefaultGameRepository.

The project has also been equipped with a scope service that allows the current repository class being used to be able to be swapped out for another from the Program.cs file. The Default Repository has currently been set as the working repository that all other classes will refer to for the time being. (This will later be replaced as the Default Repository is intended only for testing)

This branch is ready to be pulled and merged with the main branch.