Created Initial Migration and Modified Models
The initial migration of the database was created and migrated to update the database with the appropriate format.

Upon sending this migration, any inconsistencies in the data with the database format were resolved and changed appropriately. (This can be seen in changes to the classes in the Models folder) Documentation for any modified fields were also adjusted appropriately.