Get All Active Games Controller Method
Completed feature to find all actively running games (games who are still in progress) and then return them to the client.

This display shows the players involved in the game, number of moves made in that game, and a visual display of the game's progress.

This feature concludes the feature set of Endpoint 3.